refactor(pipeline): break 2 orphan cycles (12g-2)
Break the last two localized 2-node import cycles, ratcheting the
frozen cycle baseline from 7 to 5.

- PipelineBuilder <-> PipelineBuilderFactory: drop the convenience
  re-export from PipelineBuilder.ts; repoint the 15 consumers (14
  bank pipelines + 1 test) to import createPipelineBuilder from
  PipelineBuilderFactory.js directly.
- LoginFormFill <-> LoginScopeResolver: extract fillOneField and the
  IFill* option/result types into a new low-level leaf
  LoginFieldFill.ts (no back-edge to either module). ScopeResolver
  imports the leaf; FormFill re-exports it so existing consumers are
  unaffected.

Baseline re-frozen to 5. The 4 relocated framework-internal Form
symbols are allowlisted for docs-coverage (already exported pre-PR).

Validation: tsc, eslint (+canaries/prettier), lint:cycles (5), 89
affected unit tests - all green.
