+++ src/Scrapers/Pipeline/Banks/Isracard/IsracardPipeline.ts
@@ -5,7 +5,7 @@
 
 import type { ScraperOptions } from '../../../Base/Interface.js';
 import type { ILoginConfig } from '../../../Base/Interfaces/Config/LoginConfig.js';
-import { createPipelineBuilder } from '../../Core/Builder/PipelineBuilder.js';
+import { createPipelineBuilder } from '../../Core/Builder/PipelineBuilderFactory.js';
 import type { IPipelineDescriptor } from '../../Core/PipelineDescriptor.js';
 import type { Procedure } from '../../Types/Procedure.js';
